+++ client/rest-high-level/src/main/java/org/elasticsearch/client/RestHighLevelClient.java
@@ -192,6 +192,7 @@ public class RestHighLevelClient implements Closeable {
     private final IndicesClient indicesClient = new IndicesClient(this);
     private final ClusterClient clusterClient = new ClusterClient(this);
     private final SnapshotClient snapshotClient = new SnapshotClient(this);
+    private final TasksClient tasksClient = new TasksClient(this);
 
     /**
      * Creates a {@link RestHighLevelClient} given the low level {@link RestClientBuilder} that allows to build the
@@ -262,6 +263,15 @@ public class RestHighLevelClient implements Closeable {
      */
     public final SnapshotClient snapshot() {
         return snapshotClient;
+    }
+
+    /**
+     * Provides a {@link TasksClient} which can be used to access the Tasks API.
+     *
+     * See <a href="https://www.elastic.co/guide/en/elasticsearch/reference/current/tasks.html">Task Management API on elastic.co</a>
+     */
+    public final TasksClient tasks() {
+        return tasksClient;
     }
 
     /**
