+++ client/rest-high-level/src/test/java/org/elasticsearch/client/documentation/ClusterClientDocumentationIT.java
@@ -19,13 +19,8 @@
 
 package org.elasticsearch.client.documentation;
 
-import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.LatchedActionListener;
-import org.elasticsearch.action.TaskOperationFailure;
-import org.elasticsearch.action.admin.cluster.node.tasks.list.ListTasksRequest;
-import org.elasticsearch.action.admin.cluster.node.tasks.list.ListTasksResponse;
-import org.elasticsearch.action.admin.cluster.node.tasks.list.TaskGroup;
 import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest;
 import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsResponse;
 import org.elasticsearch.action.ingest.PutPipelineRequest;
@@ -39,21 +34,15 @@ import org.elasticsearch.common.unit.ByteSizeUnit;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.xcontent.XContentType;
 import org.elasticsearch.indices.recovery.RecoverySettings;
-import org.elasticsearch.tasks.TaskId;
-import org.elasticsearch.tasks.TaskInfo;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
-import static java.util.Collections.emptyList;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.greaterThanOrEqualTo;
-import static org.hamcrest.Matchers.notNullValue;
 
 /**
  * This class is used to generate the Java Cluster API documentation.
@@ -188,89 +177,6 @@ public class ClusterClientDocumentationIT extends ESRestHighLevelClientTestCase 
             // tag::put-settings-execute-async
             client.cluster().putSettingsAsync(request, listener); // <1>
             // end::put-settings-execute-async
-
-            assertTrue(latch.await(30L, TimeUnit.SECONDS));
-        }
-    }
-
-    public void testListTasks() throws IOException {
-        RestHighLevelClient client = highLevelClient();
-        {
-            // tag::list-tasks-request
-            ListTasksRequest request = new ListTasksRequest();
-            // end::list-tasks-request
-
-            // tag::list-tasks-request-filter
-            request.setActions("cluster:*"); // <1>
-            request.setNodes("nodeId1", "nodeId2"); // <2>
-            request.setParentTaskId(new TaskId("parentTaskId", 42)); // <3>
-            // end::list-tasks-request-filter
-
-            // tag::list-tasks-request-detailed
-            request.setDetailed(true); // <1>
-            // end::list-tasks-request-detailed
-
-            // tag::list-tasks-request-wait-completion
-            request.setWaitForCompletion(true); // <1>
-            request.setTimeout(TimeValue.timeValueSeconds(50)); // <2>
-            request.setTimeout("50s"); // <3>
-            // end::list-tasks-request-wait-completion
-        }
-
-        ListTasksRequest request = new ListTasksRequest();
-
-        // tag::list-tasks-execute
-        ListTasksResponse response = client.cluster().listTasks(request);
-        // end::list-tasks-execute
-
-        assertThat(response, notNullValue());
-
-        // tag::list-tasks-response-tasks
-        List<TaskInfo> tasks = response.getTasks(); // <1>
-        // end::list-tasks-response-tasks
-
-        // tag::list-tasks-response-calc
-        Map<String, List<TaskInfo>> perNodeTasks = response.getPerNodeTasks(); // <1>
-        List<TaskGroup> groups = response.getTaskGroups(); // <2>
-        // end::list-tasks-response-calc
-
-        // tag::list-tasks-response-failures
-        List<ElasticsearchException> nodeFailures = response.getNodeFailures(); // <1>
-        List<TaskOperationFailure> taskFailures = response.getTaskFailures(); // <2>
-        // end::list-tasks-response-failures
-
-        assertThat(response.getNodeFailures(), equalTo(emptyList()));
-        assertThat(response.getTaskFailures(), equalTo(emptyList()));
-        assertThat(response.getTasks().size(), greaterThanOrEqualTo(2));
-    }
-
-    public void testListTasksAsync() throws Exception {
-        RestHighLevelClient client = highLevelClient();
-        {
-            ListTasksRequest request = new ListTasksRequest();
-
-            // tag::list-tasks-execute-listener
-            ActionListener<ListTasksResponse> listener =
-                    new ActionListener<ListTasksResponse>() {
-                        @Override
-                        public void onResponse(ListTasksResponse response) {
-                            // <1>
-                        }
-
-                        @Override
-                        public void onFailure(Exception e) {
-                            // <2>
-                        }
-                    };
-            // end::list-tasks-execute-listener
-
-            // Replace the empty listener by a blocking listener in test
-            final CountDownLatch latch = new CountDownLatch(1);
-            listener = new LatchedActionListener<>(listener, latch);
-
-            // tag::list-tasks-execute-async
-            client.cluster().listTasksAsync(request, listener); // <1>
-            // end::list-tasks-execute-async
 
             assertTrue(latch.await(30L, TimeUnit.SECONDS));
         }
