+++ docs/java-rest/high-level/supported-apis.asciidoc
@@ -104,11 +104,9 @@ include::indices/put_template.asciidoc[]
 The Java High Level REST Client supports the following Cluster APIs:
 
 * <<java-rest-high-cluster-put-settings>>
-* <<java-rest-high-cluster-list-tasks>>
 * <<java-rest-high-cluster-put-pipeline>>
 
 include::cluster/put_settings.asciidoc[]
-include::cluster/list_tasks.asciidoc[]
 include::cluster/put_pipeline.asciidoc[]
 
 == Snapshot APIs
@@ -122,3 +120,11 @@ The Java High Level REST Client supports the following Snapshot APIs:
 include::snapshot/get_repository.asciidoc[]
 include::snapshot/create_repository.asciidoc[]
 include::snapshot/delete_repository.asciidoc[]
+
+== Tasks APIs
+
+The Java High Level REST Client supports the following Tasks APIs:
+
+* <<java-rest-high-tasks-list>>
+
+include::tasks/list_tasks.asciidoc[]
